requirements.txt
streamlit
pymongo
python-dotenv
pandas
streamlit-option-menu